refactor(web): moved MockClientWithOption to web
Previously we had MockClientWithOption in web_test. This moves it
to the web package to keep it together with MockClient.
