+++ web/quickfeed_helper_test.go
@@ -1,47 +1,11 @@
 package web_test
 
 import (
-	"net/http"
-	"net/http/httptest"
 	"testing"
 
-	"connectrpc.com/connect"
-	"github.com/quickfeed/quickfeed/ci"
-	"github.com/quickfeed/quickfeed/database"
-	"github.com/quickfeed/quickfeed/internal/qtest"
 	"github.com/quickfeed/quickfeed/qf"
-	"github.com/quickfeed/quickfeed/qf/qfconnect"
-	"github.com/quickfeed/quickfeed/scm"
-	"github.com/quickfeed/quickfeed/web"
 	"github.com/quickfeed/quickfeed/web/auth"
-	"github.com/quickfeed/quickfeed/web/interceptor"
 )
-
-func MockClientWithOption(t *testing.T, db database.Database, mockOpt scm.MockOption, clientOpts ...connect.ClientOption) (qfconnect.QuickFeedServiceClient, *auth.TokenManager) {
-	t.Helper()
-	mgr := scm.MockManager(t, mockOpt)
-	logger := qtest.Logger(t)
-	qfService := web.NewQuickFeedService(logger.Desugar(), db, mgr, web.BaseHookOptions{}, &ci.Local{})
-
-	tm, err := auth.NewTokenManager(db)
-	if err != nil {
-		t.Fatal(err)
-	}
-	opts := connect.WithInterceptors(
-		interceptor.NewValidationInterceptor(logger),
-		interceptor.NewTokenAuthInterceptor(logger, tm, db),
-		interceptor.NewUserInterceptor(logger, tm),
-		interceptor.NewAccessControlInterceptor(tm),
-	)
-	router := http.NewServeMux()
-	router.Handle(qfconnect.NewQuickFeedServiceHandler(qfService, opts))
-	server := httptest.NewUnstartedServer(router)
-	server.EnableHTTP2 = true
-	server.StartTLS()
-	t.Cleanup(server.Close)
-
-	return qfconnect.NewQuickFeedServiceClient(server.Client(), server.URL, clientOpts...), tm
-}
 
 func Cookie(t *testing.T, tm *auth.TokenManager, user *qf.User) string {
 	t.Helper()
